QuantumCube
Using a QuantumCube is much better. You can play Go and Othello all together. To achieve this you need this template 60 (+1) times to play with 6 Players. Try a modus with Gothello and Checkers in a group of four Players. No colored cubes needed. No more dividing into Black and White. At first, cutting and folding the cubes are time-consuming but it saves money and CO2.